Working Version.
The Case and File Apis are now based on a single base class which
handles all common operations.
