+++ utilities.py
@@ -1,0 +1,3 @@
+
+def getHistology(uuid):
+	pass
